Completed port for Legacy tests. Should move Performance tests in a separate folder. Legacy perf tests are failing.
