+++ src/System.Net.Sockets/tests/FunctionalTests/project.lock.json
@@ -619,7 +619,7 @@
           "lib/portable-net45+aspnetcore50+win+wpa81+wp80/xunit.core.dll": {}
         }
       },
-      "xunit.netcore.extensions/1.0.0-prerelease-00075": {
+      "xunit.netcore.extensions/1.0.0-prerelease-00076": {
         "dependencies": {
           "System.Diagnostics.Debug": "4.0.10-beta-23127",
           "System.IO": "4.0.10-beta-23127",
@@ -1967,12 +1967,12 @@
         "lib/portable-net45+aspnetcore50+win+wpa81+wp80/xunit.runner.utility.xml"
       ]
     },
-    "xunit.netcore.extensions/1.0.0-prerelease-00075": {
+    "xunit.netcore.extensions/1.0.0-prerelease-00076": {
       "serviceable": true,
-      "sha512": "crNU2PbilYFmbAwmiIYSGVSTRrY4Su+h4qmkeotRe41jxyJnzcfZwtY2/xFK9RiXFYjiTJrF/pRc589eNFljsw==",
+      "sha512": "3z/syoA+m4DcO9RSNVUqiYlk+GQrOcCs95bzxSdaNmZmi7IhAG5n7vElzamJpOdtt/n++qKECP60DRte4eITiA==",
       "files": [
-        "xunit.netcore.extensions.1.0.0-prerelease-00075.nupkg",
-        "xunit.netcore.extensions.1.0.0-prerelease-00075.nupkg.sha512",
+        "xunit.netcore.extensions.1.0.0-prerelease-00076.nupkg",
+        "xunit.netcore.extensions.1.0.0-prerelease-00076.nupkg.sha512",
         "xunit.netcore.extensions.nuspec",
         "lib/dotnet/Xunit.NetCore.Extensions.dll"
       ]
